Heavily refactor how zip files are mounted, which will allow mods opting-in for modifiable file systems. Also add support for modifiable overlays.
