+++ gradle.properties
@@ -5,7 +5,7 @@ group = org.quiltmc
 description = The mod loading component of Quilt
 url = https://github.com/quiltmc/quilt-loader
 # Don't forget to change this in QuiltLoaderImpl as well
-quilt_loader = 0.29.2-beta.4
+quilt_loader = 0.29.2-alexiil-modifiable-fs.1
 
 # Fabric & Quilt Libraries
 asm = 9.8
