+++ src/main/java/org/quiltmc/loader/impl/filesystem/FileSystemSource.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, 2023 QuiltMC
+ * Copyright 2025 QuiltMC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,21 +16,20 @@
 
 package org.quiltmc.loader.impl.filesystem;
 
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
+import java.io.IOException;
+
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
 
-@QuiltLoaderInternal(QuiltLoaderInternalType.LEGACY_EXPOSED)
-public final class QuiltMemoryPath extends QuiltMapPath<@NotNull QuiltMemoryFileSystem, @NotNull QuiltMemoryPath> {
+@QuiltLoaderInternal(QuiltLoaderInternalType.NEW_INTERNAL)
+interface FileSystemSource {
 
-	QuiltMemoryPath(@NotNull QuiltMemoryFileSystem fs, @Nullable QuiltMemoryPath parent, String name) {
-		super(fs, parent, name);
-	}
+	boolean needsClosing();
 
-	@Override
-	@NotNull
-	QuiltMemoryPath getThisPath() {
-		return this;
-	}
+	boolean isOpen();
+
+	void open(QuiltBaseFileSystem<?, ?> fs);
+
+	void close(QuiltBaseFileSystem<?, ?> fs) throws IOException;
+
 }
