+++ src/main/java/org/quiltmc/loader/impl/filesystem/QuiltUnifiedPath.java
@@ -16,14 +16,15 @@
 
 package org.quiltmc.loader.impl.filesystem;
 
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternal;
 import org.quiltmc.loader.impl.util.QuiltLoaderInternalType;
 
 @QuiltLoaderInternal(QuiltLoaderInternalType.NEW_INTERNAL)
-public class QuiltUnifiedPath extends QuiltMapPath<QuiltUnifiedFileSystem, QuiltUnifiedPath> {
+public class QuiltUnifiedPath extends QuiltMapPath<@NotNull QuiltUnifiedFileSystem, @NotNull QuiltUnifiedPath> {
 
-	QuiltUnifiedPath(QuiltUnifiedFileSystem fs, @Nullable QuiltUnifiedPath parent, String name) {
+	QuiltUnifiedPath(@NotNull QuiltUnifiedFileSystem fs, @Nullable QuiltUnifiedPath parent, String name) {
 		super(fs, parent, name);
 	}
