+++ src/main/java/org/quiltmc/loader/impl/util/ReadOnlyByteArrayChannel.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2025 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.quiltmc.loader.impl.util;
 
 import java.io.IOException;
@@ -41,6 +57,7 @@ public class ReadOnlyByteArrayChannel implements SeekableByteChannel {
 
 		int max = Math.min(dst.remaining(), length - position);
 		dst.put(source, offset + position, max);
+		position += max;
 		return max;
 	}
